refactor(tools): move pi tool projection out of adapter
