+++ src/backend/pi-stream-adapter.test.ts
@@ -10,12 +10,9 @@ import type {
   SimpleStreamOptions,
 } from "@earendil-works/pi-ai";
 import {
-  APPLY_PATCH_FREEFORM_DESCRIPTION,
-  APPLY_PATCH_LARK_GRAMMAR,
   PiStreamAdapter,
   type PiStreamFunction,
   stripOpenAIResponsesReplayItemIds,
-  toPiTools,
 } from "@/backend/dev/pi-stream-adapter";
 import type {
   ProviderStreamEvent,
@@ -92,69 +89,6 @@ function emptyTextBlocks(messages: Context["messages"]) {
 }
 
 describe("PiStreamAdapter", () => {
-  test("projects apply_patch as a pi custom tool with a function fallback", () => {
-    const tools = toPiTools([
-      {
-        name: "apply_patch",
-        description: "JSON schema fallback description",
-        parameters: {
-          type: "object",
-          properties: { input: { type: "string" } },
-          required: ["input"],
-          additionalProperties: false,
-        },
-      },
-    ]);
-
-    expect(tools).toEqual([
-      {
-        type: "custom",
-        name: "apply_patch",
-        description: APPLY_PATCH_FREEFORM_DESCRIPTION,
-        format: {
-          type: "grammar",
-          syntax: "lark",
-          definition: APPLY_PATCH_LARK_GRAMMAR,
-        },
-        fallback: {
-          parameters: {
-            type: "object",
-            properties: { input: { type: "string" } },
-            required: ["input"],
-            additionalProperties: false,
-          },
-        },
-      },
-    ]);
-  });
-
-  test("projects PascalCase ApplyPatch as a pi custom tool without changing execution name", () => {
-    const tools = toPiTools([
-      {
-        name: "ApplyPatch",
-        description: "JSON schema fallback description",
-        parameters: {
-          type: "object",
-          properties: { input: { type: "string" } },
-          required: ["input"],
-        },
-      },
-    ]);
-
-    expect(tools?.[0]).toMatchObject({
-      type: "custom",
-      name: "ApplyPatch",
-      description: APPLY_PATCH_FREEFORM_DESCRIPTION,
-      fallback: {
-        parameters: {
-          type: "object",
-          properties: { input: { type: "string" } },
-          required: ["input"],
-        },
-      },
-    });
-  });
-
   test("removes OpenAI Responses replay item IDs before provider submission", async () => {
     let sanitizedPayload: unknown;
     const stream: PiStreamFunction = (
